add s3 download function

+++ src/main/java/com/repring/be/MusicRepository.java
@@ -3,8 +3,10 @@ package com.repring.be;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.lang.reflect.Member;
 import java.util.List;
 
 @Repository
-public interface MusicRepository extends JpaRepository<Music, Integer> {
+public interface MusicRepository extends JpaRepository<Music, Long> {
+
 }
